requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.8
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
anyio==4.4.0
asyncpg==0.29.0
attrs==23.2.0
automat==22.10.0
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.4.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
constantly==23.10.4
contourpy==1.2.1
cryptography==43.0.0
cssselect==1.2.0
cycler==0.12.1
deep-translator==1.11.4
defusedxml==0.7.1
dnspython==2.6.1
email-validator==2.2.0
fastapi==0.111.0
fastapi-cli==0.0.4
filelock==3.15.4
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.9.0
gitdb==4.0.11
gitpython==3.1.43
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.25.1
hyperlink==21.0.0
idna==3.7
incremental==24.7.1
itemadapter==0.9.0
itemloaders==1.3.1
jinja2==3.1.4
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
langchain==0.3.1
langchain-core==0.3.8
langchain-text-splitters==0.3.0
langsmith==0.1.130
lxml==5.2.2
markdown-it-py==3.0.0
markupsafe==2.1.5
matplotlib==3.9.1
mdurl==0.1.2
multidict==6.1.0
numpy==1.26.4
orjson==3.10.6
packaging==24.1
pandas==2.2.2
parsel==1.9.1
pillow==10.4.0
protego==0.3.1
protobuf==5.27.2
psycopg2-binary==2.9.9
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1-modules==0.4.0
pycparser==2.22
pydantic==2.8.2
pydantic-core==2.20.1
pydeck==0.9.1
pydispatcher==2.0.7
pygments==2.18.0
pyopenssl==24.2.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
pyyaml==6.0.1
queuelib==1.7.0
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
rich==13.7.1
rpds-py==0.19.1
safetensors==0.4.5
schedule==1.2.2
scrapy==2.11.2
seaborn==0.13.2
service-identity==24.1.0
setuptools==70.3.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
sqlalchemy==2.0.31
starlette==0.37.2
streamlit==1.36.0
tenacity==8.5.0
tldextract==5.1.2
tokenizers==0.20.0
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.5
transformers==4.45.1
twisted==24.3.0
twisted-iocpsupport==1.0.4
typer==0.12.3
typing-extensions==4.12.2
ujson==5.10.0
urllib3==2.2.2
uvicorn==0.30.1
w3lib==2.2.1
watchdog==4.0.1
watchfiles==0.22.0
websockets==12.0
yarl==1.13.1
zope-interface==6.4.post2